Move dialogue-related patches into their own file
