+++ MWSE/PatchDialogue.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace mwse::patch::dialogue {
+	void install();
+}
